backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
pydantic==2.10.6
pydantic-settings==2.8.0
alembic==1.15.1
PyJWT==2.10.1
python-dotenv==1.0.1
httpx==0.28.1
passlib==1.7.4
bcrypt==3.2.2